+++ FinanceBackEnd/tests/Finance.Application.Tests/Services/CreditCards/CreditCardServiceTestsDeleteOwner.cs
@@ -1,0 +1,37 @@
+using CQRSDispatch;
+using Finance.Application.Commands.CreditCards;
+using Microsoft.AspNetCore.Http;
+
+namespace Finance.Application.Tests.Services.CreditCards;
+
+public partial class CreditCardServiceTests : IDisposable
+{
+    [Fact]
+    public async Task DeleteOwner_DispatchesDeleteOwnerCommandWithCorrectId()
+    {
+        var resourceId = Guid.NewGuid();
+
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteCreditCardOwnerCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Success());
+
+        await _sut.DeleteOwner(resourceId);
+
+        _dispatcher.Verify(d => d.DispatchAsync<CommandResult>(
+            It.Is<DeleteCreditCardOwnerCommand>(c => c.EntityId == resourceId),
+            It.IsAny<HttpRequest?>()),
+            Times.Once);
+    }
+
+    [Fact]
+    public async Task DeleteOwner_ReturnsDispatchResult()
+    {
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteCreditCardOwnerCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Success());
+
+        var result = await _sut.DeleteOwner(Guid.NewGuid());
+
+        Assert.True(result.IsSuccess);
+    }
+}
