+++ FinanceBackEnd/tests/Finance.Application.Tests/Services/DebitOrigins/DebitOriginServiceTests.Delete.cs
@@ -1,0 +1,79 @@
+using CQRSDispatch;
+using Finance.Application.Commands.DebitOrigins;
+using Finance.Application.Services.DebitOrigins;
+using Microsoft.AspNetCore.Http;
+
+namespace Finance.Application.Tests.Services.DebitOrigins;
+
+public partial class DebitOriginServiceTests : IDisposable
+{
+    [Fact]
+    public async Task Delete_WhenDispatchSucceeds_ReturnsSuccess()
+    {
+        var request = new DeleteDebitOriginRequest([Guid.NewGuid()]);
+
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteDebitOriginCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Success());
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteDebitOriginOwnerCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Success());
+
+        var result = await _sut.Delete(request);
+
+        Assert.True(result.IsSuccess);
+    }
+
+    [Fact]
+    public async Task Delete_DeletesOwnerForEachId()
+    {
+        var ids = new[] { Guid.NewGuid(), Guid.NewGuid() };
+        var request = new DeleteDebitOriginRequest(ids);
+
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteDebitOriginCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Success());
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteDebitOriginOwnerCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Success());
+
+        await _sut.Delete(request);
+
+        _dispatcher.Verify(d => d.DispatchAsync<CommandResult>(
+            It.IsAny<DeleteDebitOriginOwnerCommand>(),
+            It.IsAny<HttpRequest?>()),
+            Times.Exactly(ids.Length));
+    }
+
+    [Fact]
+    public async Task Delete_WhenDeleteDispatchFails_ReturnsFailure()
+    {
+        var request = new DeleteDebitOriginRequest([Guid.NewGuid()]);
+
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteDebitOriginCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Failure("delete error"));
+
+        var result = await _sut.Delete(request);
+
+        Assert.False(result.IsSuccess);
+        Assert.Equal("Type DebitOrigin delete operation failed", result.ErrorMessage);
+    }
+
+    [Fact]
+    public async Task Delete_WhenDeleteDispatchFails_DoesNotDispatchOwnerDelete()
+    {
+        var request = new DeleteDebitOriginRequest([Guid.NewGuid()]);
+
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteDebitOriginCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Failure("delete error"));
+
+        await _sut.Delete(request);
+
+        _dispatcher.Verify(d => d.DispatchAsync<CommandResult>(
+            It.IsAny<DeleteDebitOriginOwnerCommand>(),
+            It.IsAny<HttpRequest?>()),
+            Times.Never);
+    }
+}
