+++ FinanceBackEnd/tests/Finance.Application.Tests/Services/DebitOrigins/DebitOriginServiceTests.SetOwner.cs
@@ -1,0 +1,29 @@
+using CQRSDispatch;
+using Finance.Application.Commands.DebitOrigins;
+using Finance.Domain.Models.Auth;
+using FinanceBackEnd.Finance.Domain.Enums;
+using Microsoft.AspNetCore.Http;
+
+namespace Finance.Application.Tests.Services.DebitOrigins;
+
+public partial class DebitOriginServiceTests : IDisposable
+{
+    [Fact]
+    public async Task SetOwner_DispatchesCreatePermissionsCommandWithOwnerLevel()
+    {
+        var resourceId = Guid.NewGuid();
+
+        _dispatcher
+            .Setup(d => d.DispatchAsync<DataResult<DebitOriginPermissions>>(It.IsAny<CreateDebitOriginPermissionsCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(DataResult<DebitOriginPermissions>.Success(new DebitOriginPermissions()));
+
+        await _sut.SetOwner(resourceId);
+
+        _dispatcher.Verify(d => d.DispatchAsync<DataResult<DebitOriginPermissions>>(
+            It.Is<CreateDebitOriginPermissionsCommand>(c =>
+                c.ResourceId == resourceId &&
+                c.PermissionLevels.Contains(PermissionLevelEnum.Owner)),
+            It.IsAny<HttpRequest?>()),
+            Times.Once);
+    }
+}
