+++ FinanceBackEnd/tests/Finance.Application.Tests/Services/Debits/DebitServiceTests.DeleteOwner.cs
@@ -1,0 +1,25 @@
+using CQRSDispatch;
+using Finance.Application.Commands.Debits;
+using Microsoft.AspNetCore.Http;
+
+namespace Finance.Application.Tests.Services.Debits;
+
+public partial class DebitServiceTests : IDisposable
+{
+    [Fact]
+    public async Task DeleteOwner_DispatchesDeleteOwnerCommandWithCorrectId()
+    {
+        var resourceId = Guid.NewGuid();
+
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteDebitOwnerCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Success());
+
+        await _sut.DeleteOwner(resourceId);
+
+        _dispatcher.Verify(d => d.DispatchAsync<CommandResult>(
+            It.Is<DeleteDebitOwnerCommand>(c => c.EntityId == resourceId),
+            It.IsAny<HttpRequest?>()),
+            Times.Once);
+    }
+}
