+++ FinanceBackEnd/tests/Finance.Application.Tests/Services/Debits/DebitServiceTests.SetOwner.cs
@@ -1,0 +1,29 @@
+using CQRSDispatch;
+using Finance.Application.Commands.Debits;
+using Finance.Domain.Models.Auth;
+using FinanceBackEnd.Finance.Domain.Enums;
+using Microsoft.AspNetCore.Http;
+
+namespace Finance.Application.Tests.Services.Debits;
+
+public partial class DebitServiceTests : IDisposable
+{
+    [Fact]
+    public async Task SetOwner_DispatchesCreatePermissionsCommandWithOwnerLevel()
+    {
+        var resourceId = Guid.NewGuid();
+
+        _dispatcher
+            .Setup(d => d.DispatchAsync<DataResult<DebitPermissions>>(It.IsAny<CreateDebitPermissionsCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(DataResult<DebitPermissions>.Success(new DebitPermissions()));
+
+        await _sut.SetOwner(resourceId);
+
+        _dispatcher.Verify(d => d.DispatchAsync<DataResult<DebitPermissions>>(
+            It.Is<CreateDebitPermissionsCommand>(c =>
+                c.ResourceId == resourceId &&
+                c.PermissionLevels.Contains(PermissionLevelEnum.Owner)),
+            It.IsAny<HttpRequest?>()),
+            Times.Once);
+    }
+}
