+++ FinanceBackEnd/tests/Finance.Application.Tests/Services/Funds/FundServiceTests.DeleteOwner.cs
@@ -1,0 +1,37 @@
+using CQRSDispatch;
+using Finance.Application.Commands;
+using Microsoft.AspNetCore.Http;
+
+namespace Finance.Application.Tests.Services.Funds;
+
+public partial class FundServiceTests : IDisposable
+{
+    [Fact]
+    public async Task DeleteOwner_DispatchesDeleteOwnerCommandWithCorrectEntityId()
+    {
+        var resourceId = Guid.NewGuid();
+
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteFundOwnerCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Success());
+
+        await _sut.DeleteOwner(resourceId);
+
+        _dispatcher.Verify(d => d.DispatchAsync<CommandResult>(
+            It.Is<DeleteFundOwnerCommand>(c => c.EntityId == resourceId),
+            It.IsAny<HttpRequest?>()),
+            Times.Once);
+    }
+
+    [Fact]
+    public async Task DeleteOwner_ReturnsDispatcherResult()
+    {
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteFundOwnerCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Success());
+
+        var result = await _sut.DeleteOwner(Guid.NewGuid());
+
+        Assert.True(result.IsSuccess);
+    }
+}
