+++ FinanceBackEnd/tests/Finance.Application.Tests/Services/Subscriptions/SubscriptionServiceTests.DeleteOwner.cs
@@ -1,0 +1,37 @@
+using CQRSDispatch;
+using Finance.Application.Commands.Subscriptions;
+using Microsoft.AspNetCore.Http;
+
+namespace Finance.Application.Tests.Services.Subscriptions;
+
+public partial class SubscriptionServiceTests : IDisposable
+{
+    [Fact]
+    public async Task DeleteOwner_DispatchesDeleteOwnerCommandWithCorrectEntityId()
+    {
+        var resourceId = Guid.NewGuid();
+
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteSubscriptionOwnerCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Success());
+
+        await _sut.DeleteOwner(resourceId);
+
+        _dispatcher.Verify(d => d.DispatchAsync<CommandResult>(
+            It.Is<DeleteSubscriptionOwnerCommand>(c => c.EntityId == resourceId),
+            It.IsAny<HttpRequest?>()),
+            Times.Once);
+    }
+
+    [Fact]
+    public async Task DeleteOwner_ReturnsDispatcherResult()
+    {
+        _dispatcher
+            .Setup(d => d.DispatchAsync<CommandResult>(It.IsAny<DeleteSubscriptionOwnerCommand>(), It.IsAny<HttpRequest?>()))
+            .ReturnsAsync(CommandResult.Success());
+
+        var result = await _sut.DeleteOwner(Guid.NewGuid());
+
+        Assert.True(result.IsSuccess);
+    }
+}
